Ensure Agent ID creation is idempotent.
The Opsmatic Agent ID creation only happened once due
to the 'refreshonly' tag. This removes that, and re-executes
the creation appropriately if any of the /identity/ files
are found to be missing.

Additionally, do not even create the Exec[] resource if
we're not installing the service, otherwise it will fail
during the uninstall of the service.